# fix(client): Concatenate string instead of using `fmt.Sprintf`